refactor(r4a): consolidate wifi under internal/wifi (was internal/canopy)
R4a descriptive-rename + reorg (per PHASE3_RECONCILE_PROPOSAL.md + the agreed
capability-first convention). The botanical "canopy" grouping held the Wi-Fi
scanner/manager and the site-survey feature; move them where they belong:

  internal/canopy/wifi    -> internal/wifi
  internal/canopy/survey  -> internal/wifi/survey
  internal/canopy/        -> removed (empty)

Pure move + import-path rewrite (24 wifi + 20 survey importers across api +
discovery/wifi_bridge); package names unchanged (wifi, survey). No logic change.

Behavior gate: golden HTTP suite + Wi-Fi/survey route tests unchanged. go build
./... + GOOS=windows green; go test ./... green; golangci-lint 0.

Convention (agreed): capability-first flat layout — drop the inconsistent
internal/{services,modules,adapters} umbrellas. Next R4a moves: internal/services
diagnostics -> internal/diagnostics; services/discovery -> internal/discovery;
services/shell/guestaudit -> internal/security; modules/harvest -> internal/
reporting (repo port stays, impl lives with reporting); then api grouping renames.
