+++ internal/api/handlers_survey.go
@@ -12,10 +12,10 @@ import (
 	"net/http"
 	"slices"
 
-	"github.com/krisarmstrong/seed/internal/canopy/survey"
 	"github.com/krisarmstrong/seed/internal/i18n"
 	"github.com/krisarmstrong/seed/internal/logging"
 	"github.com/krisarmstrong/seed/internal/validation"
+	"github.com/krisarmstrong/seed/internal/wifi/survey"
 )
 
 // Survey validation limits.
