+++ internal/api/handlers_survey_analysis.go
@@ -8,9 +8,9 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/krisarmstrong/seed/internal/canopy/survey"
 	"github.com/krisarmstrong/seed/internal/i18n"
 	"github.com/krisarmstrong/seed/internal/logging"
+	"github.com/krisarmstrong/seed/internal/wifi/survey"
 )
 
 // getSurveyDeadZones handles GET /api/survey/dead-zones?id=xxx&threshold=-75.
