+++ internal/api/handlers_survey_floorplan.go
@@ -10,10 +10,10 @@ import (
 	"slices"
 	"strings"
 
-	"github.com/krisarmstrong/seed/internal/canopy/survey"
 	"github.com/krisarmstrong/seed/internal/i18n"
 	"github.com/krisarmstrong/seed/internal/logging"
 	"github.com/krisarmstrong/seed/internal/validation"
+	"github.com/krisarmstrong/seed/internal/wifi/survey"
 )
 
 // UpdateFloorPlanRequest contains floor plan image and dimension parameters.
