+++ internal/api/handlers_survey_floors_data.go
@@ -7,9 +7,9 @@ package api
 import (
 	"net/http"
 
-	"github.com/krisarmstrong/seed/internal/canopy/survey"
 	"github.com/krisarmstrong/seed/internal/i18n"
 	"github.com/krisarmstrong/seed/internal/logging"
+	"github.com/krisarmstrong/seed/internal/wifi/survey"
 )
 
 // SetActiveFloorRequest contains parameters for setting the active floor.
