+++ internal/api/handlers_survey_license_internal_test.go
@@ -5,7 +5,7 @@ package api
 import (
 	"testing"
 
-	"github.com/krisarmstrong/seed/internal/canopy/survey"
+	"github.com/krisarmstrong/seed/internal/wifi/survey"
 )
 
 // newRoamFixture returns a survey with one Active + one Passive sample
