+++ internal/api/services.go
@@ -3,8 +3,6 @@ package api
 import (
 	"github.com/krisarmstrong/seed/internal/alerts"
 	"github.com/krisarmstrong/seed/internal/auth"
-	"github.com/krisarmstrong/seed/internal/canopy/survey"
-	"github.com/krisarmstrong/seed/internal/canopy/wifi"
 	"github.com/krisarmstrong/seed/internal/database"
 	"github.com/krisarmstrong/seed/internal/dhcp"
 	"github.com/krisarmstrong/seed/internal/engine"
@@ -26,6 +24,8 @@ import (
 	"github.com/krisarmstrong/seed/internal/services/vlan"
 	"github.com/krisarmstrong/seed/internal/timeseries/retention"
 	"github.com/krisarmstrong/seed/internal/update"
+	"github.com/krisarmstrong/seed/internal/wifi"
+	"github.com/krisarmstrong/seed/internal/wifi/survey"
 )
 
 // ServiceContainer holds all application services organized by domain.
