+++ internal/wifi/doc.go
