+++ internal/wifi/export_test.go
