+++ internal/wifi/scanner.go
