+++ internal/wifi/scanner_darwin_extended_test.go
@@ -5,7 +5,7 @@ package wifi_test
 import (
 	"testing"
 
-	"github.com/krisarmstrong/seed/internal/canopy/wifi"
+	"github.com/krisarmstrong/seed/internal/wifi"
 )
 
 // TestParseAirportLineSecurityTypes tests parsing various security type formats.
