+++ internal/wifi/scanner_linux.go
