+++ internal/wifi/survey/airmapper_parser.go
