+++ internal/wifi/survey/analysis.go
