+++ internal/wifi/survey/colorscale.go
