+++ internal/wifi/survey/interpolation.go
