+++ internal/wifi/survey/migration_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/krisarmstrong/seed/internal/canopy/survey"
+	"github.com/krisarmstrong/seed/internal/wifi/survey"
 )
 
 func TestMigrateToMultiFloor_AlreadyMigrated(t *testing.T) {
